Creation of bottom_bar
Created a new index file and moved the code of the original index file (now home_screen) to the home_screen file.
Renamed the route homeScreen to Initial Screen  (that is located in the index file) so we can call the method of the BottomBar.
The bottom_bar file calls the home_screen and profile_screen in order to be visible